Add `indent` to `node.position`
This update adds support for a new property on `position`,
next to the currently available `start` and `end`, which,
on multiline elements, contains at which column the node
started on each line.

This information is useful for tools which need to know
at which columns, for example a blockquote, starts.
This is neccesary because markdown allows lazy indentation,
such as:

```markdown
>   This is a paragraph
continuing on the next line
```

...where the second line has a different indent from
others.