Remove `instanceof` operator in `file`
File returns its given options if they are itself a file. This
makes it easy to patch custom properties relating to files.
However, the `instaceof` broke this for different versions of
mdast, so if a plug-in uses a different version than the host
package, this resulted in loss of patched properties.