Refactor to externalise core API
This refactor externalises the API (process/ use / parse /
transform / compile). The two reasons for that:

*   Now **retext** can use a similar API, without duplicate
    code;

*   Less commits relating to API design will be in this
    repository.

This does change the exports object of `lib/parse` and
`lib/compile`, in that they previously exposed different
interfaces (a function with a constructor on top). This
was normalized into just a constructor (`Parser` /
`Compiler`) which receive a virtual file and settings,
and expose respectivly either a `parse` or a `compile`
method.

This additionally changes some internals (where plug-ins)
are stored, and removes the limitation of not allowing
multiple equal attachers.