Refactor support for entities
Previously, entities were decoded among their surrounding characters
in text nodes.  This produced some caveats, which can be seen in
wooorm/atom-linter-alex#4.

This update adds decoded entities as their own separate text nodes, and
additionally adds support for warning messages relating to errors in
these entities.

For example, `&copy` is valid HTML (and Markdown) but produces a
Parse Error, which in **mdast** results in a message on the VFile.

See wooorm/parse-entities for more information.

Closes GH-66.
Closes wooorm/atom-linter-alex#4