Add support for syntax tree input to remark(1)
This update removes the previously supported `-a, --ast` flag, and
renames it to `--tree-out`.  In addition, `--tree-in` is now supported
to depict that the read input is an already parsed syntax tree in JSON.
And, this commit introduces a new `-t, --tree` flag which turns both
`--tree-in` and `--tree-out` on, thus specifying a tree input and a
tree output.

Closes GH-126.
Closes GH-156.

Reviewed-By: Eugene Sharygin <eush77@gmail.com>