Add support for syntax tree input to remark(1)
This update removes the previously supported `-a, --ast` flag, and
renames it to `--tree-out`.  In addition, `--tree-in` is now supported
to depict that the input read from stdin(4) is an already parsed
syntax tree in JSON.  And, this commit introduces a new `-t, --tree`
flag which turns both `--tree-in` and `--tree-out` on, thus specifying
a tree input on stdin(4) and a tree output on stdout(4).

Closes GH-126.