Remove `position: false` code
The interface still works as expected: `position: false` turns
off `position` objects in the generated syntax tree (useful for
debugging).  However, the code to enable this was removed, instead
opting for a post-processing step which strips `position` objects
at the end of `Parser#parse`.

The `position: false` behaviour is still tested.

Closes GH-146.