Fix list-item continuation and fenced-code
Previously, fenced code was part of the previous list-item, even if it
was not indented.  This is now changed: for fenced code to exist in
a list-item it must be indented, just like thematic breaks.

This fix ensures the parser functions like GitHub (RedCarpet),
markdown-it, and CommonMark.

If, instead, you want to allow lazy continuation for fenced code
blocks, use `pedantic: true` mode.

Closes GH-169.