Fix non-magic globs from matching all files
For a while, remark(1) included all files nested in directories
if that directory was given on the CLI.  For example, if `content`
was given, `content/foo.jpg` was included by mistake.  Now, if a glob
does not contain magic (as in, it is a file-path) and represents a
directory, only files with known markdown extensions in it are
included.

Related to https://gitter.im/wooorm/remark?at=570b1b9905e328bb386d0d99.