Remove false-positive in incorrect stdin(4) detection
Previously, when the file-descriptor `0` (stdin) was a FIFO, and
the process reported not being in a TTY, and files were given, an
incorrect error was thrown.

Now, when files are given, stdin(4) is completely and silently ignored,
when no files are given and the process reports being a TTY, an error
is thrown, and when no files are given and the process reports not
being a TTY, stdin(4) is read.

Closes GH-176.
Closes denysdovhan/gulp-remark#9.