Fix bug when installing both CLI and API as siblings
Previously, there was a race conditional where, when `remark-cli` was
installed before `remark`, an error was throw that `remark-cli` could
not be found.

This commit fixes that by checking, in `remark/cli` whether
`remark-cli` is installed next to it, in which case it’s loaded
and no warning is triggered.

Closes GH-206.