remark-stringify: add custom `stringLength` measuring
Previously, when table cells were padded to align them, each cell’s
value (`s`) was measures in size from the string’s `.length` property.
That’s of course lacking: it doesn’t work well with tabs, CJK, or
emoji.  However, it’s not really a solvable problem, as you could
view the markdown source on different OSes, editors, terminals, or
using different fonts, and note that they each could display
characters at different sizes.

To “solve” this, it’s now possible to pass in a `stringLength`
function, and return the size it should be percieved as, to
`remark-stringify`.

Closes GH-200.