Remove support for `yaml` from core
Previously, yaml was support out of the box in core.  As YAML
isn’t the only frontmatter available, and adding more frontmatters
to core is superfluous, and in lieu of keeping core small, this
behaviour is removed here and will be moved to another plugin:
remark-frontmatter.

Related to wooorm/remark-lint#149.