Fix: pedantic remark-stringify can produce markdown that cannot be pedantically parsed
The specific problem this PR addresses is with underscores in emphasis.
In pedantic mode, underscores in emphasis are only allowed if the
emphasis is created with asterisks, not underscores:

-  `*alpha_bravo*` compiles to `<em>alpha_bravo</em>`
-  `_alpha_bravo_` compiles to `<em>alpha</em>bravo_`

Unfortunately, `remark-stringify` as it currently stands will stringify
both inputs to `_alpha_bravo`, meaning that the result of the chain:

    remark-parser -> remark-stringify -> remark-parser -> remark-to-rehype -> rehype-stringify

will be different from the result of the chain:

    remark-parser -> remark-to-rehype -> rehype-stringify

in pedantic mode.

The goal of this PR is to ensure that those two chains have identical
outputs; or, put another way, to ensure that `remark-stringify` is not a
destructive process, even in pedantic mode.