This PR fixes inline code in table cells.
Previously, inline code got preference over the table cell itself.
This was added due to GH-73 in GH-83 four years ago.

That behaviour does not match current GFM and lead to unexpected results in
GH-295, GH-320, and GH-396.

**This PR fixes them, but introduces a conundrum: GFM does not allow empty
inline code in cells. Which is weird. Should we support that (why not? 🤷) or
match GFM?**

Related to GH-73.
Related to GH-83.
Closes GH-295.
Closes GH-320.
Closes GH-396.