Fix numbered lists interrupting paragraphs
In ancient times, Markdown was a bit weird, most of that was solved in CM.
One such example is that here we don’t start
2. …a list, but here we do
1. …the first one is paragraph continuation text, the second starts with a `1`,
which does turn it into a new list.

Previous behavior was mixed, but as GFM now follows CM, and pedantic is out of
favor, this commit uses the single CM way instead of three separate equally
weird modes.